+++ src/pages/background/handler/userScheduleHandler.ts
@@ -7,7 +7,7 @@ import renameSchedule from '@pages/background/lib/renameSchedule';
 import switchSchedule from '@pages/background/lib/switchSchedule';
 import type { UserScheduleMessages } from '@shared/messages/UserScheduleMessages';
 import { Course } from '@shared/types/Course';
-import type { MessageHandler } from 'chrome-extension-toolkit';
+import type { MessageHandler, Serialized } from 'chrome-extension-toolkit';
 
 const userScheduleHandler: MessageHandler<UserScheduleMessages> = {
     addCourse({ data, sendResponse }) {
@@ -30,6 +30,9 @@ const userScheduleHandler: MessageHandler<UserScheduleMessages> = {
     },
     renameSchedule({ data, sendResponse }) {
         renameSchedule(data.scheduleId, data.newName).then(sendResponse);
+    },
+    addCourseByURL({ data, sendResponse }): Promise<void> | void {
+        throw new Error('Function not implemented.');
     },
 };
