+++ src/shared/messages/UserScheduleMessages.ts
@@ -44,4 +44,11 @@ export interface UserScheduleMessages {
      * @returns undefined if successful, otherwise an error message
      */
     renameSchedule: (data: { scheduleId: string; newName: string }) => string | undefined;
+
+    /**
+     * Adds a course by URL
+     * @param data
+     * @returns
+     */
+    addCourseByURL: (data: { url: string; method: string; body?: string; response: 'json' | 'text' }) => unknown;
 }
